move engine and gamestate to own directory
