+++ src/engine/GameState.ts
@@ -1,9 +1,9 @@
 import type { GameState, BoardCard, BoardSkill, Player } from "./Engine.ts";
 
 // Import JSON types generated by quicktype
-import type { Enchantments, Card, Cards } from "./types/cardTypes.ts";
-import type { Tier } from "./types/shared.ts";
-import type { EncounterDays, Group } from "./types/encounterTypes.ts";
+import type { Enchantments, Card, Cards } from "../types/cardTypes.ts";
+import type { Tier } from "../types/shared.ts";
+import type { EncounterDays, Group } from "../types/encounterTypes.ts";
 
 /* Bugs
